separate model and db logic for tasks, add task update tests
