+++ lib/db/schemas/tasks.js
@@ -8,6 +8,10 @@ const dynamodb = require('../dynamodb_client');
  * Module
  */
 
+const ATTRIBUTE_NAME_SUBSTITUTIONS = {
+    name: '#NAME'
+};
+
 const schema = {
     TableName : 'Tasks',
     KeySchema: [       
@@ -24,8 +28,58 @@ const schema = {
     }
 };
 
+function toDynamoDBFromPartialJSON(userId, taskId, data) {
+    // TODO: merge with toDynamoDBFromJSON method
+    const payload = {
+        userId: {
+            S: userId
+        },
+        id: {
+            S: taskId
+        }
+    };
+    const providedKeys = Object.keys(data);
+    if (providedKeys.includes('name')) {
+        payload.name = { S: data.name };
+    }
+    if (providedKeys.includes('frequency')) {
+        payload.frequency = { N: String(data.frequency) };
+    }
+    if (providedKeys.includes('completionDates')) {
+        payload.completionDates = {
+            L: data.completionDates.map(date => ({ S: date }))
+        };
+    }
+    if (providedKeys.includes('archived')) {
+        payload.archived = { BOOL: data.archived };
+    }
+    return payload;
+}
 
-function fromDynamoDb(data) {
+function toDynamoDBFromJSON(data) {
+    return {
+        id: {
+            S: data.id
+        },
+        userId: {
+            S: data.userId
+        },
+        name: {
+            S: data.name
+        },
+        frequency: {
+            N: String(data.frequency)
+        },
+        completionDates: {
+            L: data.completionDates.map(date => ({ S: date }))
+        },
+        archived: {
+            BOOL: data.archived
+        }
+    };
+}
+
+function toJSONfromDynamoDB(data) {
     const expectedKeys = ['id', 'userId', 'name', 'frequency', 'completionDates', 'archived'];
     expectedKeys.forEach((key) => {
         if (!Object.keys(data).includes(key)) {
@@ -64,7 +118,7 @@ async function getTaskById(userId, taskId) {
         error.httpStatusCode = 404;
         throw error;
     }
-    return fromDynamoDb(response.Item);
+    return toJSONfromDynamoDB(response.Item);
 }
 
 
@@ -79,7 +133,7 @@ async function listTasksForUser(userId) {
         Select: 'ALL_ATTRIBUTES',
         KeyConditionExpression: 'userId = :v1'
     }).promise();
-    return response.Items.map(item => fromDynamoDb(item));
+    return response.Items.map(item => toJSONfromDynamoDB(item));
 }
 
 
@@ -109,14 +163,73 @@ async function removeAll() {
     });
 }
 
+async function put(data) {
+    const payload = toDynamoDBFromJSON(data);
+    await dynamodb.putItem({
+        TableName: 'Tasks',
+        Item: payload
+    }).promise();
+}
+
+async function update(userId, taskId, data) {
+    const payload = toDynamoDBFromPartialJSON(userId, taskId, data);
+
+    // assign key-value pairs to attribute value identifiers, and
+    // map reserved attribute names to substitutions
+    const expressionAttributeNames = {};
+    const expressionMapping = Object.entries(payload)
+        .filter(([attribute]) => !['id', 'userId'].includes(attribute))
+        .map(([attribute, value], index) => {
+            let attributeName = attribute;
+            if (Object.keys(ATTRIBUTE_NAME_SUBSTITUTIONS).includes(attribute)) {
+                attributeName = ATTRIBUTE_NAME_SUBSTITUTIONS[attribute];
+                expressionAttributeNames[attributeName] = attribute;
+            }
+            return {
+                expressionKey: `:v${index}`,
+                attribute: attributeName,
+                value
+            };
+        });
+
+    // build update body
+    const expressionAttributeValues = expressionMapping.reduce((acc, mapping) => {
+        acc[mapping.expressionKey] = mapping.value;
+        return acc;
+    }, {});
+    const updateExpressionParts = expressionMapping
+        .map(mapping => `${mapping.attribute}=${mapping.expressionKey}`);
+    const updateExpression = `SET ${updateExpressionParts.join(', ')}`;
+
+    const updateItemPayload = {
+        TableName: 'Tasks',
+        Key: {
+            id: payload.id,
+            userId: payload.userId
+        },
+        ExpressionAttributeValues: expressionAttributeValues,
+        UpdateExpression: updateExpression
+    };
+
+    if (Object.keys(expressionAttributeNames).length > 0) {
+        updateItemPayload.ExpressionAttributeNames = expressionAttributeNames;
+    }
+
+    await dynamodb.updateItem(updateItemPayload).promise();
+}
+
 /**
  * Module exports
  */
 
 module.exports = {
     schema,
-    fromDynamoDb,
+    put,
     getTaskById,
     listTasksForUser,
     removeAll,
+    toJSONfromDynamoDB,
+    toDynamoDBFromPartialJSON,
+    toDynamoDBFromJSON,
+    update,
 };
