+++ lib/db/schemas/users.js
@@ -22,7 +22,7 @@ const schema = {
     }
 };
 
-function toDynamoDB(data) {
+function toDynamoDBFromJSON(data) {
     return {
         id: {
             S: data.id
@@ -36,7 +36,7 @@ function toDynamoDB(data) {
     };
 }
 
-function fromDynamoDb(data) {
+function toJSONfromDynamoDB(data) {
     const expectedKeys = ['id', 'name', 'created'];
     expectedKeys.forEach((key) => {
         if (!Object.keys(data).includes(key)) {
@@ -95,11 +95,11 @@ async function getById(userId) {
         error.httpStatusCode = 404;
         throw error;
     }
-    return fromDynamoDb(response.Item);
+    return toJSONfromDynamoDB(response.Item);
 }
 
 async function create(data) {
-    const payload = toDynamoDB(data);
+    const payload = toDynamoDBFromJSON(data);
     await dynamodb.putItem({
         TableName: 'Users',
         Item: payload
@@ -113,7 +113,7 @@ async function create(data) {
 module.exports = {
     schema,
     create,
-    fromDynamoDb,
+    toJSONfromDynamoDB,
     getById,
     removeAll
 };
