+++ lib/models/task.js
@@ -3,17 +3,12 @@
  */
 
 const uuid = require('uuid');
-const dynamodb = require('../db/dynamodb_client');
 const schemas = require('../db/schemas');
 const {queueAsyncFunctions} = require('../util');
 
 /**
  * Module
  */
-
-const ATTRIBUTE_NAME_SUBSTITUTIONS = {
-    name: '#NAME'
-};
 
 class TaskData {
     constructor(id, userId, name, frequency, completionDates, archived) {
@@ -27,57 +22,6 @@ class TaskData {
 
     static create(userId, name, frequency) {
         return new TaskData(uuid.v4(), userId, name, frequency, [], false);
-    }
-
-    static partialToDynamoDB(userId, taskId, data) {
-        // TODO: add helper methods to dynamodb client, use in forEach
-        const payload = {
-            userId: {
-                S: userId
-            },
-            id: {
-                S: taskId
-            }
-        };
-        const providedKeys = Object.keys(data);
-        if (providedKeys.includes('name')) {
-            payload.name = { S: data.name };
-        }
-        if (providedKeys.includes('frequency')) {
-            payload.frequency = { N: String(data.frequency) };
-        }
-        if (providedKeys.includes('completionDates')) {
-            payload.completionDates = {
-                L: data.completionDates.map(date => ({ S: date }))
-            };
-        }
-        if (providedKeys.includes('archived')) {
-            payload.archived = { BOOL: data.archived };
-        }
-        return payload;
-    }
-
-    toDynamoDB() {
-        return {
-            id: {
-                S: this.id
-            },
-            userId: {
-                S: this.userId
-            },
-            name: {
-                S: this.name
-            },
-            frequency: {
-                N: String(this.frequency)
-            },
-            completionDates: {
-                L: this.completionDates.map(date => ({ S: date }))
-            },
-            archived: {
-                BOOL: this.archived
-            }
-        };
     }
 
     static fromJSON(data) {
@@ -124,61 +68,14 @@ class Task {
     static async create(userId, data) {
         // TODO: validate userId in authorization
         const taskData = TaskData.create(userId, data.name, data.frequency);
-        const payload = taskData.toDynamoDB();
-        await dynamodb.putItem({
-            TableName: 'Tasks',
-            Item: payload
-        }).promise();
+        await schemas.tasks.put(taskData.toJSON());
         // TODO: handle errors
         return taskData;
     }
 
     static async update(userId, taskId, data) {
         // TODO: validate userId in authorization
-        const payload = TaskData.partialToDynamoDB(userId, taskId, data);
-
-        // assign key-value pairs to attribute value identifiers, and
-        // map reserved attribute names to substitutions
-        const expressionAttributeNames = {};
-        const expressionMapping = Object.entries(payload)
-            .filter(([attribute]) => !['id', 'userId'].includes(attribute))
-            .map(([attribute, value], index) => {
-                let attributeName = attribute;
-                if (Object.keys(ATTRIBUTE_NAME_SUBSTITUTIONS).includes(attribute)) {
-                    attributeName = ATTRIBUTE_NAME_SUBSTITUTIONS[attribute];
-                    expressionAttributeNames[attributeName] = attribute;
-                }
-                return {
-                    expressionKey: `:v${index}`,
-                    attribute: attributeName,
-                    value
-                };
-            });
-
-        // build update body
-        const expressionAttributeValues = expressionMapping.reduce((acc, mapping) => {
-            acc[mapping.expressionKey] = mapping.value;
-            return acc;
-        }, {});
-        const updateExpressionParts = expressionMapping
-            .map(mapping => `${mapping.attribute}=${mapping.expressionKey}`);
-        const updateExpression = `SET ${updateExpressionParts.join(', ')}`;
-
-        const updateItemPayload = {
-            TableName: 'Tasks',
-            Key: {
-                id: payload.id,
-                userId: payload.userId
-            },
-            ExpressionAttributeValues: expressionAttributeValues,
-            UpdateExpression: updateExpression
-        };
-
-        if (Object.keys(expressionAttributeNames).length > 0) {
-            updateItemPayload.ExpressionAttributeNames = expressionAttributeNames;
-        }
-
-        await dynamodb.updateItem(updateItemPayload).promise();
+        await schemas.tasks.update(userId, taskId, data);
         // TODO: handle errors
         const taskData = await this.getById(userId, taskId);
         return taskData;
