+++ lib/routes/task.js
@@ -85,7 +85,7 @@ async function handleUpdateTask(req, res, next) {
     const taskId = req.params.id;
     const userId = 'lukeskywalker';
     try {
-        const task = await models.task.Task.put(userId, taskId, req.body);
+        const task = await models.task.Task.update(userId, taskId, req.body);
         const responseData = task.toJSON();
         logResponse(requestId, 200, req);
         res.json(responseData);
